use `card` class to render list of posts
I actually don't much care for this style in this case. It also seems to
break up the list in `w3m` to make things more complicated. It may be
useful in other contexts, though. Perhaps if I use it on other sites
I'll name the template `deck.html` because the representation is more a
deck of cards than a list of elements.